requirements.txt
Flask==3.0.0
Pillow==10.1.0
transformers==4.35.0
torch==2.1.0
torchvision>=0.16.0
numpy>=1.24.0
gTTS==2.4.0
werkzeug==3.0.1
kagglehub>=0.3.0
matplotlib>=3.7.0

